pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "finance-dashboard"
version = "0.1.0"
description = "Finance dashboard application"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.40.0",
    "pandas>=2.2.3",
    "numpy>=2.1.3",
    "plotly>=5.24.1",
    "altair>=5.4.1",
    "scipy>=1.13.1",
    "pyarrow>=17.0.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.10.6",
    "pyyaml>=6.0",
]

[tool.setuptools.packages.find]
where = ["src"]


[tool.setuptools.package-data]
"*" = ["*.yml", "*.yaml"]

[project.scripts]
fetch-data = "finance_dashboard.fetch_data:main"
dev-run-app = "finance_dashboard.run_app:main"
run-app = "finance_dashboard.run_app:main_prod"

